Add generic coordinator interpolator

+++ src/main/kotlin/io/github/yeyu/easing/NoEase1D.kt
@@ -4,5 +4,5 @@ import io.github.yeyu.easing.interpolator.GenericInterpolator1D
 
 class NoEase1D<T: Number>(from: T, to: T) : Ease<T> {
     override val interpolator = GenericInterpolator1D(from, to)
-    override fun next(at: Float): T = interpolator.next(at)
+    override fun next(at: Double): T = interpolator.next(at)
 }
