+++ src/main/kotlin/io/github/yeyu/easing/player/ReversingEasePlayer.kt
@@ -19,7 +19,7 @@ open class ReversingEasePlayer<T : Number>(
         }
 
     override fun next(): T {
-        val selectFrame = if (currentFrame == numberOfFrames - 2) 1.0f else (currentFrame + 1).toFloat() / (numberOfFrames - 1)
+        val selectFrame = if (currentFrame == numberOfFrames - 2) 1.0 else (currentFrame + 1).toDouble() / (numberOfFrames - 1)
         current = easeFn.next(selectFrame)
         if (reverse) currentFrame-- else currentFrame++
         if (currentFrame == numberOfFrames - 1) reverse = true
